requirements.txt
huggingface_hub==0.30.2
matplotlib==3.10.1
pandas==2.2.3
plotly==5.24.1
protobuf==6.30.2
PyPDF2==3.0.1
python_docx==1.1.2
reportlab==4.3.1
streamlit==1.44.1
